📝 타입 코드 정리
